+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.4] - 2023-03-19
+
+### Added
+- Add `SqlTark\Connection\OracleConnection` for PDO_OCI driver
+- Add `SqlTark\Connection\PostgresConnection` for PDO_PGSQL driver
+
+### Changed
+- Change `SqlTark\Connection\Config` implementation
+
 ## [0.9.3] - 2023-03-16
 
 ### Changed